+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+
+- Promotional video btn in the above the fold of an active event
+
 ## [0.0.68] - 2024-09-19
 
 ### Fix